requirements.txt
boto3
flask
pycryptodome
pyjwt
Padding